+++ src/sys/os/linux/syscalls/consts/mod.rs
@@ -6,7 +6,7 @@
 
 #![allow(non_camel_case_types, clippy::upper_case_acronyms)]
 
-crate::CONST! { SYS_DOC = "Linux `sys/syscall.h` constants for the current compilation target."; }
+crate::CONST! { _DOC_SYS = "Linux `sys/syscall.h` constants for the current compilation target."; }
 
 crate::items! {
     #[cfg(target_arch = "aarch64")] mod aarch64;
@@ -16,19 +16,19 @@ crate::items! {
     #[cfg(target_arch = "x86_64")] mod x86_64;
 }
 
-#[doc = SYS_DOC!()]
+#[doc = _DOC_SYS!()]
 #[cfg(target_arch = "aarch64")]
 pub type LINUX_SYS = aarch64::LINUX_SYS_AARCH64;
-#[doc = SYS_DOC!()]
+#[doc = _DOC_SYS!()]
 #[cfg(target_arch = "arm")]
 pub type LINUX_SYS = arm::LINUX_SYS_ARM;
-#[doc = SYS_DOC!()]
+#[doc = _DOC_SYS!()]
 #[cfg(any_target_arch_riscv)]
 pub type LINUX_SYS = riscv::LINUX_SYS_RISCV;
-#[doc = SYS_DOC!()]
+#[doc = _DOC_SYS!()]
 #[cfg(target_arch = "x86")]
 pub type LINUX_SYS = x86::LINUX_SYS_X86;
-#[doc = SYS_DOC!()]
+#[doc = _DOC_SYS!()]
 #[cfg(target_arch = "x86_64")]
 pub type LINUX_SYS = x86_64::LINUX_SYS_X86_64;